fix test_load_checkpoints
