+++ tests/mome/test_mome_checkpoints.py
@@ -8,7 +8,7 @@ from vllm.mome.models import MoMEModel
 from vllm.model_executor.models.llama import LlamaForCausalLM
 
 def test_load_checkpoints():
-    supported_mome_modules = LlamaForCausalLM.supported_lora_modules
+    supported_mome_modules = LlamaForCausalLM.supported_mome_modules
     packed_modules_mapping = LlamaForCausalLM.packed_modules_mapping
     expected_mome_modules: List[str] = []
     for module in supported_mome_modules:
